Refactor: UserDTO 및 Service 분리
